Remove Utility Functions from Public API
Fixes https://github.com/ReactiveX/RxJava/issues/1797
